Move TF tests with process sets into separate scripts
