+++ test/parallel/test_tensorflow.py
@@ -29,91 +29,26 @@ import pytest
 import sys
 import tensorflow as tf
 from horovod.tensorflow.util import _executing_eagerly
-from tensorflow.python.framework import ops
 from tensorflow.python.ops import resource_variable_ops
 from tensorflow.python.ops import variables as tf_ops_variables
-
-import warnings
-
-import horovod.tensorflow as hvd
 
 sys.path.append(os.path.join(os.path.dirname(__file__), os.pardir, 'utils'))
 
 from common import mpi_env_rank_and_size, skip_or_fail_gpu_test
 
-if hasattr(tf, 'ConfigProto'):
-    config = tf.ConfigProto()
-    config.gpu_options.allow_growth = True
+import horovod.tensorflow as hvd
 
-if hasattr(tf, 'config') and hasattr(tf.config, 'experimental') \
-        and hasattr(tf.config.experimental, 'set_memory_growth'):
-    gpus = tf.config.experimental.list_physical_devices('GPU')
-    for gpu in gpus:
-        tf.config.experimental.set_memory_growth(gpu, True)
-else:
-    # Specifies the config to use with eager execution. Does not preclude
-    # tests from running in the graph mode.
-    tf.enable_eager_execution(config=config)
-
-ccl_supported_types = set([tf.uint8, tf.int8, tf.uint16, tf.int16,
-                           tf.int32, tf.int64, tf.float32, tf.float64])
+from base_test_tensorflow import *
 
 _IS_TF2 = LooseVersion(tf.__version__) >= LooseVersion('2.0.0')
 _is_mac = platform.system() == 'Darwin'
 
-# Set environment variable to enable adding/removing process sets after initializing Horovod.
-os.environ["HOROVOD_DYNAMIC_PROCESS_SETS"] = "1"
-
-
-class TensorFlowTests(tf.test.TestCase):
+class TensorFlowTests(BaseTensorFlowTests):
     """
     Tests for ops in horovod.tensorflow.
     """
-
     def __init__(self, *args, **kwargs):
         super(TensorFlowTests, self).__init__(*args, **kwargs)
-        warnings.simplefilter('module')
-        if hasattr(tf, 'contrib') and hasattr(tf.contrib, 'eager'):
-            self.tfe = tf.contrib.eager
-        else:
-            self.tfe = tf
-
-    def evaluate(self, tensors):
-        if _executing_eagerly():
-            return self._eval_helper(tensors)
-        sess = ops.get_default_session()
-        if sess is None:
-            with self.test_session(config=config) as sess:
-                return sess.run(tensors)
-        else:
-            return sess.run(tensors)
-
-    def assign(self, variables, values):
-        if _executing_eagerly():
-            for var, val in zip(variables, values):
-                var.assign(val)
-        else:
-            sess = ops.get_default_session()
-            if sess is None:
-                with self.test_session(config=config) as sess:
-                    for var, val in zip(variables, values):
-                        var.load(val, sess)
-            else:
-                for var, val in zip(variables, values):
-                    var.load(val, sess)
-
-    def random_uniform(self, *args, **kwargs):
-        if hasattr(tf, 'random') and hasattr(tf.random, 'set_seed'):
-            tf.random.set_seed(1234)
-            return tf.random.uniform(*args, **kwargs)
-        else:
-            tf.set_random_seed(1234)
-            return tf.random_uniform(*args, **kwargs)
-
-    def filter_supported_types(self, types):
-        if 'CCL_ROOT' in os.environ:
-           types = [t for t in types if t in ccl_supported_types]
-        return types
 
     def test_gpu_required(self):
         if not tf.test.is_gpu_available(cuda_only=True):
@@ -170,41 +105,6 @@ class TensorFlowTests(tf.test.TestCase):
         size = self.evaluate(hvd.size_op())
         self.assertTrue(size == hvd.size(),
                         "hvd.size_op produces incorrect results")
-
-    def test_horovod_size_op_process_set(self):
-        """Test that the size returned by hvd.size_op(process_set_id) is correct."""
-        hvd.init()
-
-        # This test does not apply if there is only one worker.
-        if hvd.size() == 1:
-            self.skipTest("Only one worker available")
-
-        single_set = hvd.add_process_set([0])
-
-        size = self.evaluate(hvd.size_op(process_set_id=single_set.process_set_id))
-        self.assertEqual(size, single_set.size(),
-                        "hvd.size_op produces incorrect results for a process set")
-
-        hvd.remove_process_set(single_set)
-
-    def test_horovod_process_set_included_op(self):
-        """Test that the result of hvd.process_set_included_op(process_set_id) is correct."""
-        hvd.init()
-
-        # This test does not apply if there is only one worker.
-        if hvd.size() == 1:
-            self.skipTest("Only one worker available")
-
-        single_set = hvd.add_process_set([0])
-
-        included = self.evaluate(hvd.process_set_included_op(process_set_id=single_set.process_set_id))
-
-        if hvd.rank() == 0:
-            self.assertEqual(included, 1)
-        else:
-            self.assertEqual(included, 0)
-
-        hvd.remove_process_set(single_set)
 
     def test_horovod_local_size_op(self):
         """Test that the local size returned by hvd.local_size_op() is correct."""
@@ -383,54 +283,6 @@ class TensorFlowTests(tf.test.TestCase):
             diff = self.evaluate(max_difference)
             self.assertTrue(diff <= threshold,
                             "hvd.allreduce produces incorrect results")
-
-    def test_horovod_allreduce_cpu_process_sets(self):
-        """ Test on CPU that allreduce correctly sums if restricted to non-global process sets"""
-        hvd.init()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        if hvd.ccl_built():
-            self.skipTest("Multiple process sets currently do not support CCL.")
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        dtypes = self.filter_supported_types([tf.int32, tf.int64, tf.float16, tf.float32, tf.float64])
-        dims = [1, 2, 3]
-        for dtype, dim in itertools.product(dtypes, dims):
-            with tf.device("/cpu:0"):
-                even_rank_tensor = self.random_uniform([17] * dim, -100, 100, dtype=dtype)
-                odd_rank_tensor = self.random_uniform([17] * dim, -100, 100, dtype=dtype)
-                if rank in even_ranks:
-                    summed = hvd.allreduce(even_rank_tensor, average=False, process_set=even_set)
-                    multiplied = even_rank_tensor * len(even_ranks)
-                if rank in odd_ranks:
-                    summed = hvd.allreduce(odd_rank_tensor, average=False, process_set=odd_set)
-                    multiplied = odd_rank_tensor * len(odd_ranks)
-                max_difference = tf.reduce_max(tf.abs(summed - multiplied))
-
-            # Threshold for floating point equality depends on number of
-            # ranks, since we're comparing against precise multiplication.
-            max_process_set_size = max(len(even_ranks), len(odd_ranks))
-            if max_process_set_size <= 3 or dtype in [tf.int32, tf.int64]:
-                threshold = 0
-            elif max_process_set_size < 10:
-                threshold = 1e-4
-            elif max_process_set_size < 15:
-                threshold = 5e-4
-            else:
-                self.skipTest("Horovod cluster too large for precise multiplication comparison")
-
-            diff = self.evaluate(max_difference)
-            self.assertTrue(diff <= threshold, "hvd.allreduce produces incorrect results")
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
 
     def test_horovod_allreduce_gpu(self):
         """Test that the allreduce works on GPUs."""
@@ -700,59 +552,6 @@ class TensorFlowTests(tf.test.TestCase):
             self.assertTrue(diff <= threshold,
                             "hvd.allreduce produces incorrect results")
 
-    def test_horovod_allreduce_gpu_process_sets(self):
-        """ Test on GPU that allreduce correctly sums if restricted to non-global process sets"""
-        # Only do this test if there are GPUs available.
-        if not tf.test.is_gpu_available(cuda_only=True):
-            self.skipTest(("No GPUs available"))
-
-        if int(os.environ.get('HOROVOD_MIXED_INSTALL', 0)):
-            # Skip if compiled with CUDA but without HOROVOD_GPU_OPERATIONS.
-            self.skipTest("Not compiled with HOROVOD_GPU_OPERATIONS")
-
-        hvd.init()
-        local_rank = hvd.local_rank()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        dtypes = [tf.int32, tf.int64, tf.float16, tf.float32, tf.float64]
-        dims = [1, 2, 3]
-        for dtype, dim in itertools.product(dtypes, dims):
-            with tf.device("/gpu:%d" % local_rank):
-                even_rank_tensor = self.random_uniform([17] * dim, -100, 100, dtype=dtype)
-                odd_rank_tensor = self.random_uniform([17] * dim, -100, 100, dtype=dtype)
-                if rank in even_ranks:
-                    summed = hvd.allreduce(even_rank_tensor, average=False, process_set=even_set)
-                    multiplied = even_rank_tensor * len(even_ranks)
-                if rank in odd_ranks:
-                    summed = hvd.allreduce(odd_rank_tensor, average=False, process_set=odd_set)
-                    multiplied = odd_rank_tensor * len(odd_ranks)
-                max_difference = tf.reduce_max(tf.abs(summed - multiplied))
-
-            # Threshold for floating point equality depends on number of
-            # ranks, since we're comparing against precise multiplication.
-            max_process_set_size = max(len(even_ranks), len(odd_ranks))
-            if max_process_set_size <= 3 or dtype in [tf.int32, tf.int64]:
-                threshold = 0
-            elif max_process_set_size < 10:
-                threshold = 1e-4
-            elif max_process_set_size < 15:
-                threshold = 5e-4
-            else:
-                self.skipTest("Horovod cluster too large for precise multiplication comparison")
-
-            diff = self.evaluate(max_difference)
-            self.assertTrue(diff <= threshold, "hvd.allreduce produces incorrect results")
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
     def test_horovod_allreduce_error(self):
         """Test that the allreduce raises an error if different ranks try to
         send tensors of different rank or dimension."""
@@ -796,38 +595,6 @@ class TensorFlowTests(tf.test.TestCase):
                          dtype=tf.int32 if rank % 2 == 0 else tf.float32)
         with self.assertRaises(tf.errors.FailedPreconditionError):
             self.evaluate(hvd.allreduce(tensor))
-
-    def test_horovod_allreduce_process_set_id_error(self):
-        """Test that allreduce raises an error if an invalid process set id
-        is specified."""
-        hvd.init()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        # This test does not apply if there is only one worker.
-        if size == 1:
-            self.skipTest("Only one worker available")
-
-        single_set = hvd.add_process_set([0])
-        rest_set = hvd.add_process_set(range(1, size))
-
-        try:
-            with tf.device("/cpu:0"):
-                tensor = tf.ones(4)
-                if rank == 0:
-                    with self.assertRaises(tf.errors.InvalidArgumentError):
-                        self.evaluate(hvd.allreduce(tensor, process_set=rest_set))
-                else:
-                    with self.assertRaises(tf.errors.InvalidArgumentError):
-                        self.evaluate(hvd.allreduce(tensor, process_set=single_set))
-                with self.assertRaises(ValueError):
-                    fake_set = hvd.ProcessSet([0])
-                    fake_set.process_set_id = 10  # you should not do this
-                    self.evaluate(hvd.allreduce(tensor, process_set=fake_set))
-        finally:
-            hvd.remove_process_set(rest_set)
-            hvd.remove_process_set(single_set)
-
 
     def test_horovod_allreduce_cpu_gpu_error(self):
         """Test that the allreduce raises an error if different ranks try to
@@ -923,72 +690,6 @@ class TensorFlowTests(tf.test.TestCase):
             self.assertLess(err, 0.00000001,
                             "gradient %s differs from expected %s, "
                             "error: %s" % (grad_out, expected, str(err)))
-
-    def test_horovod_allreduce_grad_cpu_process_sets(self):
-        """Test the correctness of the allreduce gradient on CPU if restricted to non-global process sets."""
-        hvd.init()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        if hvd.ccl_built():
-            self.skipTest("Multiple process sets currently do not support CCL.")
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        # As of TensorFlow v1.9, gradients are not supported on
-        # integer tensors
-        dtypes = [tf.float32, tf.float64]
-        dims = [1, 2, 3]
-        for dtype, dim in itertools.product(dtypes, dims):
-            with tf.device("/cpu:0"):
-                if _executing_eagerly():
-                    even_rank_tensor = self.tfe.Variable(self.random_uniform(
-                        [5] * dim, -100, 100, dtype=dtype))
-                    odd_rank_tensor = self.tfe.Variable(self.random_uniform(
-                        [5] * dim, -100, 100, dtype=dtype))
-                    with tf.GradientTape() as tape:
-                        if rank in even_ranks:
-                            summed = hvd.allreduce(even_rank_tensor, average=False,
-                                                   process_set=even_set)
-                        elif rank in odd_ranks:
-                            summed = hvd.allreduce(odd_rank_tensor, average=False,
-                                                   process_set=odd_set)
-                else:
-                    even_rank_tensor = self.random_uniform([5] * dim, -100, 100, dtype=dtype)
-                    odd_rank_tensor = self.random_uniform([5] * dim, -100, 100, dtype=dtype)
-                    if rank in even_ranks:
-                        summed = hvd.allreduce(even_rank_tensor, average=False,
-                                               process_set=even_set)
-                    elif rank in odd_ranks:
-                        summed = hvd.allreduce(odd_rank_tensor, average=False,
-                                               process_set=odd_set)
-
-                if rank in even_ranks:
-                    tensor = even_rank_tensor
-                    set_size = len(even_ranks)
-                elif rank in odd_ranks:
-                    tensor = odd_rank_tensor
-                    set_size = len(odd_ranks)
-
-                grad_ys = tf.ones([5] * dim)
-                if _executing_eagerly():
-                    grad_out = tape.gradient(summed, tensor, grad_ys)
-                else:
-                    grad = tf.gradients(summed, tensor, grad_ys)[0]
-                    grad_out = self.evaluate(grad)
-
-            expected = np.ones([5] * dim) * set_size
-            err = np.linalg.norm(expected - grad_out)
-            self.assertLess(err, 0.00000001,
-                            "gradient %s differs from expected %s, "
-                            "error: %s" % (grad_out, expected, str(err)))
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
 
     def test_horovod_allreduce_grad_gpu(self):
         """Test the correctness of the allreduce gradient on GPU."""
@@ -1217,179 +918,6 @@ class TensorFlowTests(tf.test.TestCase):
                 self.assertLess(err, 0.00000001,
                                 "gradient %s differs from expected %s, "
                                 "error: %s" % (grad_out, expected, str(err)))
-
-    def test_horovod_grouped_allreduce_cpu_process_sets(self):
-        """Test on CPU that the grouped allreduce correctly sums if restricted to non-global process sets"""
-        hvd.init()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        if hvd.ccl_built():
-            self.skipTest("Multiple process sets currently do not support CCL.")
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        dtypes = self.filter_supported_types([tf.int32, tf.int64, tf.float16, tf.float32, tf.float64])
-        dims = [1, 2, 3]
-        for dtype, dim in itertools.product(dtypes, dims):
-            with tf.device("/cpu:0"):
-                even_rank_tensors = [self.random_uniform(
-                    [17] * dim, -100, 100, dtype=dtype) for _ in range(5)]
-                odd_rank_tensors = [self.random_uniform(
-                    [17] * dim, -100, 100, dtype=dtype) for _ in range(5)]
-                if rank in even_ranks:
-                    summed = hvd.grouped_allreduce(even_rank_tensors, average=False, process_set=even_set)
-                    multiplied = [tensor * len(even_ranks) for tensor in even_rank_tensors]
-                elif rank in odd_ranks:
-                    summed = hvd.grouped_allreduce(odd_rank_tensors, average=False, process_set=odd_set)
-                    multiplied = [tensor * len(odd_ranks) for tensor in odd_rank_tensors]
-            max_difference = tf.reduce_max([tf.reduce_max(tf.abs(t1 - t2)) for t1, t2 in zip(summed, multiplied)])
-
-            # Threshold for floating point equality depends on number of
-            # ranks, since we're comparing against precise multiplication.
-            max_process_set_size = max(len(even_ranks), len(odd_ranks))
-            if max_process_set_size <= 3 or dtype in [tf.int32, tf.int64]:
-                threshold = 0
-            elif max_process_set_size < 10:
-                threshold = 1e-4
-            elif max_process_set_size < 15:
-                threshold = 5e-4
-            else:
-                self.skipTest("Horovod cluster too large for precise multiplication comparison")
-
-            diff = self.evaluate(max_difference)
-            self.assertTrue(diff <= threshold, "hvd.grouped_allreduce produces incorrect results")
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
-
-    def test_horovod_grouped_allreduce_gpu_process_sets(self):
-        """Test on GPU that the grouped allreduce correctly sums if restricted to non-global process sets"""
-        if not tf.test.is_gpu_available(cuda_only=True):
-            self.skipTest("No GPUs available")
-        if int(os.environ.get('HOROVOD_MIXED_INSTALL', 0)):
-            # Skip if compiled with CUDA but without HOROVOD_GPU_OPERATIONS.
-            self.skipTest("Not compiled with HOROVOD_GPU_OPERATIONS")
-        hvd.init()
-        rank = hvd.rank()
-        local_rank = hvd.local_rank()
-        size = hvd.size()
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        dtypes = self.filter_supported_types([tf.int32, tf.int64, tf.float16, tf.float32, tf.float64])
-        dims = [1, 2, 3]
-        for dtype, dim in itertools.product(dtypes, dims):
-            with tf.device("/gpu:%d" % local_rank):
-                even_rank_tensors = [self.random_uniform(
-                    [17] * dim, -100, 100, dtype=dtype) for _ in range(5)]
-                odd_rank_tensors = [self.random_uniform(
-                    [17] * dim, -100, 100, dtype=dtype) for _ in range(5)]
-                if rank in even_ranks:
-                    summed = hvd.grouped_allreduce(even_rank_tensors, average=False, process_set=even_set)
-                    multiplied = [tensor * len(even_ranks) for tensor in even_rank_tensors]
-                elif rank in odd_ranks:
-                    summed = hvd.grouped_allreduce(odd_rank_tensors, average=False, process_set=odd_set)
-                    multiplied = [tensor * len(odd_ranks) for tensor in odd_rank_tensors]
-            max_difference = tf.reduce_max([tf.reduce_max(tf.abs(t1 - t2)) for t1, t2 in zip(summed, multiplied)])
-
-            # Threshold for floating point equality depends on number of
-            # ranks, since we're comparing against precise multiplication.
-            max_process_set_size = max(len(even_ranks), len(odd_ranks))
-            if max_process_set_size <= 3 or dtype in [tf.int32, tf.int64]:
-                threshold = 0
-            elif max_process_set_size < 10:
-                threshold = 1e-4
-            elif max_process_set_size < 15:
-                threshold = 5e-4
-            else:
-                self.skipTest("Horovod cluster too large for precise multiplication comparison")
-
-            diff = self.evaluate(max_difference)
-            self.assertTrue(diff <= threshold, "hvd.grouped_allreduce produces incorrect results")
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
-    def test_horovod_grouped_allreduce_grad_cpu_process_sets(self):
-        """Test the correctness of the grouped allreduce gradient on CPU 
-        if restricted to non-global process sets."""
-        hvd.init()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        if hvd.ccl_built():
-            self.skipTest("Multiple process sets currently do not support CCL.")
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        # As of TensorFlow v1.9, gradients are not supported on
-        # integer tensors
-        dtypes = [tf.float32, tf.float64]
-        dims = [1, 2, 3]
-        for dtype, dim in itertools.product(dtypes, dims):
-            with tf.device("/cpu:0"):
-                if _executing_eagerly():
-                    even_rank_tensors = [self.tfe.Variable(self.random_uniform(
-                        [5] * dim, -100, 100, dtype=dtype)) for _ in range(5)]
-                    odd_rank_tensors = [self.tfe.Variable(self.random_uniform(
-                        [5] * dim, -100, 100, dtype=dtype)) for _ in range(5)]
-                    with tf.GradientTape(persistent=True) as tape:
-                        if rank in even_ranks:
-                            summed = hvd.grouped_allreduce(even_rank_tensors, average=False,
-                                                           process_set=even_set)
-                        elif rank in odd_ranks:
-                            summed = hvd.grouped_allreduce(odd_rank_tensors, average=False,
-                                                           process_set=odd_set)
-                else:
-                    even_rank_tensors = [self.random_uniform(
-                        [5] * dim, -100, 100, dtype=dtype) for _ in range(5)]
-                    odd_rank_tensors = [self.random_uniform(
-                        [5] * dim, -100, 100, dtype=dtype) for _ in range(5)]
-                    if rank in even_ranks:
-                        summed = hvd.grouped_allreduce(even_rank_tensors, average=False,
-                                                       process_set=even_set)
-                    elif rank in odd_ranks:
-                        summed = hvd.grouped_allreduce(odd_rank_tensors, average=False,
-                                                       process_set=odd_set)
-
-                if rank in even_ranks:
-                    tensors = even_rank_tensors
-                    set_size = len(even_ranks)
-                elif rank in odd_ranks:
-                    tensors = odd_rank_tensors
-                    set_size = len(odd_ranks)
-
-                grads_ys = [tf.ones([5] * dim, dtype=dtype) for _ in range(5)]
-                if _executing_eagerly():
-                    grads_out = [tape.gradient(s, t, g) for s, t, g in zip(summed, tensors, grads_ys)]
-                else:
-                    grads = [tf.gradients(s, t, g)[0] for s, t, g in zip(summed, tensors, grads_ys)]
-                    grads_out = [self.evaluate(grad) for grad in grads]
-
-            expected = np.ones([5] * dim) * set_size
-            for grad_out in grads_out:
-                err = np.linalg.norm(expected - grad_out)
-                self.assertLess(err, 0.00000001,
-                                "gradient %s differs from expected %s, "
-                                "error: %s" % (grad_out, expected, str(err)))
-                
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
 
     def test_horovod_allgather_cpu(self):
         """Test that the allgather correctly gathers 1D, 2D, 3D tensors."""
@@ -1957,207 +1485,6 @@ class TensorFlowTests(tf.test.TestCase):
                             "error: %s" %
                             (grad_out, expected, str(err)))
 
-    def test_horovod_allgather_cpu_process_sets(self):
-        """Test that the allgather correctly gathers 1D, 2D, 3D tensors if restricted to non-global process sets."""
-        hvd.init()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        if hvd.ccl_built():
-            self.skipTest("Multiple process sets currently do not support CCL.")
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        if rank in even_ranks:
-            set_size = len(even_ranks)
-            set_ranks = even_ranks
-            this_set = even_set
-        elif rank in odd_ranks:
-            set_size = len(odd_ranks)
-            set_ranks = odd_ranks
-            this_set = odd_set
-
-        dtypes = [tf.uint8, tf.int8, tf.uint16, tf.int16,
-                  tf.int32, tf.int64, tf.float16, tf.float32,
-                  tf.float64, tf.bool]
-        dims = [1, 2, 3]
-        for dtype, dim in itertools.product(dtypes, dims):
-            tensor = tf.ones([17] * dim) * rank
-            if dtype == tf.bool:
-                tensor = tensor % 2
-            tensor = tf.cast(tensor, dtype=dtype)
-            with tf.device("/cpu:0"):
-                gathered = hvd.allgather(tensor, process_set=this_set)
-
-            gathered_tensor = self.evaluate(gathered)
-            self.assertEqual(list(gathered_tensor.shape),
-                             [17 * set_size] + [17] * (dim - 1))
-
-            for i in range(set_size):
-                rank_tensor = tf.slice(gathered_tensor,
-                                       [i * 17] + [0] * (dim - 1),
-                                       [17] + [-1] * (dim - 1))
-                self.assertEqual(list(rank_tensor.shape), [17] * dim)
-                # tf.equal() does not support tf.uint16 as of TensorFlow 1.2,
-                # so need to cast rank_tensor to tf.int32.
-                if dtype != tf.bool:
-                    value = set_ranks[i]
-                else:
-                    value = set_ranks[i] % 2
-                self.assertTrue(
-                    self.evaluate(tf.reduce_all(
-                        tf.equal(tf.cast(rank_tensor, tf.int32), value))),
-                    "hvd.allgather produces incorrect gathered tensor")
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
-
-    def test_horovod_allgather_gpu_process_sets(self):
-        """Test that the allgather correctly gathers 1D, 2D, 3D tensors if restricted to non-global process sets."""
-
-        if not tf.test.is_gpu_available(cuda_only=True):
-            self.skipTest("No GPUs available")
-
-        if int(os.environ.get('HOROVOD_MIXED_INSTALL', 0)):
-            # Skip if compiled with CUDA but without HOROVOD_GPU_OPERATIONS.
-            self.skipTest("Not compiled with HOROVOD_GPU_OPERATIONS")
-
-        hvd.init()
-        rank = hvd.rank()
-        local_rank = hvd.local_rank()
-        size = hvd.size()
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        if rank in even_ranks:
-            set_size = len(even_ranks)
-            set_ranks = even_ranks
-            this_set = even_set
-        elif rank in odd_ranks:
-            set_size = len(odd_ranks)
-            set_ranks = odd_ranks
-            this_set = odd_set
-
-        dtypes = [tf.uint8, tf.int8, tf.uint16, tf.int16,
-                  tf.int32, tf.int64, tf.float16, tf.float32,
-                  tf.float64, tf.bool]
-        dims = [1, 2, 3]
-        for dtype, dim in itertools.product(dtypes, dims):
-            tensor = tf.ones([17] * dim) * rank
-            if dtype == tf.bool:
-                tensor = tensor % 2
-            tensor = tf.cast(tensor, dtype=dtype)
-            with tf.device("/gpu:%d" % local_rank):
-                gathered = hvd.allgather(tensor, process_set=this_set)
-
-            gathered_tensor = self.evaluate(gathered)
-            self.assertEqual(list(gathered_tensor.shape),
-                             [17 * set_size] + [17] * (dim - 1))
-
-            for i in range(set_size):
-                rank_tensor = tf.slice(gathered_tensor,
-                                       [i * 17] + [0] * (dim - 1),
-                                       [17] + [-1] * (dim - 1))
-                self.assertEqual(list(rank_tensor.shape), [17] * dim)
-                # tf.equal() does not support tf.uint16 as of TensorFlow 1.2,
-                # so need to cast rank_tensor to tf.int32.
-                if dtype != tf.bool:
-                    value = set_ranks[i]
-                else:
-                    value = set_ranks[i] % 2
-                self.assertTrue(
-                    self.evaluate(tf.reduce_all(
-                        tf.equal(tf.cast(rank_tensor, tf.int32), value))),
-                    "hvd.allgather produces incorrect gathered tensor")
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
-
-    def test_horovod_allgather_grad_cpu_process_sets(self):
-        """Test the correctness of the allgather gradient on CPU if restricted to non-global process sets."""
-        hvd.init()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        if hvd.ccl_built():
-            self.skipTest("Multiple process sets currently do not support CCL.")
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        if rank in even_ranks:
-            set_ranks = even_ranks
-            this_set = even_set
-        elif rank in odd_ranks:
-            set_ranks = odd_ranks
-            this_set = odd_set
-
-        # As of TensorFlow v1.9, gradients are not supported on
-        # integer tensors
-        dtypes = [tf.float32, tf.float64]
-        dims = [1, 2, 3]
-        for dtype, dim in itertools.product(dtypes, dims):
-            tensor_sizes = [3, 2, 7, 4, 6, 8, 10] * 5
-            tensor_sizes = tensor_sizes[:size]
-            set_tensor_sizes = [tensor_sizes[rk] for rk in set_ranks]
-
-            with tf.device("/cpu:0"):
-                if _executing_eagerly():
-                    with tf.GradientTape() as tape:
-                        tensor = self.tfe.Variable(
-                            tf.ones([tensor_sizes[rank]] + [17] * (dim - 1)) * rank)
-                        if dtype == tf.bool:
-                            tensor = tensor % 2
-                        tensor = tf.cast(tensor, dtype=dtype)
-                        gathered = hvd.allgather(tensor, process_set=this_set)
-                        grad_list = []
-                        for r, tensor_size in zip(set_ranks, set_tensor_sizes):
-                            g = tf.ones([tensor_size] + [17] * (dim - 1)) * r
-                            grad_list.append(g)
-                        grad_ys = tf.concat(grad_list, axis=0)
-                    grad_out = tape.gradient(gathered, tensor, grad_ys)
-                else:
-                    tensor = tf.ones([tensor_sizes[rank]] + [17] * (dim - 1)) * rank
-                    if dtype == tf.bool:
-                        tensor = tensor % 2
-                    tensor = tf.cast(tensor, dtype=dtype)
-                    gathered = hvd.allgather(tensor, process_set=this_set)
-
-                    grad_list = []
-                    for r, tensor_size in zip(set_ranks, set_tensor_sizes):
-                        g = tf.ones([tensor_size] + [17] * (dim - 1)) * r
-                        grad_list.append(g)
-                    grad_ys = tf.concat(grad_list, axis=0)
-
-                    grad = tf.gradients(gathered, tensor, grad_ys)[0]
-                    grad_out = self.evaluate(grad)
-
-            expected = np.ones(
-                [tensor_sizes[rank]] + [17] * (dim - 1)
-            ) * rank
-            err = np.linalg.norm(expected - grad_out)
-            self.assertLess(err, 0.00000001,
-                            "gradient %s differs from expected %s, "
-                            "error: %s" %
-                            (grad_out, expected, str(err)))
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
-
     def test_horovod_broadcast_cpu(self):
         """Test that the broadcast correctly broadcasts 1D, 2D, 3D tensors on CPU."""
         hvd.init()
@@ -2384,158 +1711,6 @@ class TensorFlowTests(tf.test.TestCase):
                                 tf.cast(root_tensor, tf.int32), tf.cast(broadcasted_tensor, tf.int32)))),
                             "Inplace hvd.broadcast_ produces incorrect broadcasted variable value")
 
-    def test_horovod_broadcast_cpu_process_sets(self):
-        """Test that the broadcast correctly broadcasts 1D, 2D, 3D tensors on CPU
-         if restricted to non-global process sets"""
-        hvd.init()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        if hvd.ccl_built():
-            self.skipTest("Multiple process sets currently do not support CCL.")
-
-        # This test does not apply if there is only one worker.
-        if size == 1:
-            self.skipTest("Only one worker available")
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        if rank in even_ranks:
-            set_size = len(even_ranks)
-            set_ranks = even_ranks
-            this_set = even_set
-        elif rank in odd_ranks:
-            set_size = len(odd_ranks)
-            set_ranks = odd_ranks
-            this_set = odd_set
-
-        dtypes = [tf.uint8, tf.int8, tf.uint16, tf.int16,
-                  tf.int32, tf.int64, tf.float16, tf.float32,
-                  tf.float64, tf.bool]
-        dims = [1, 2, 3]
-        root_ranks = list(set_ranks)
-        for dtype, dim, root_rank in itertools.product(dtypes, dims, root_ranks):
-            with tf.device("/cpu:0"):
-                tensor = tf.ones([17] * dim) * rank
-                root_tensor = tf.ones([17] * dim) * root_rank
-                if dtype == tf.bool:
-                    tensor = tensor % 2
-                    root_tensor = root_tensor % 2
-                tensor = tf.cast(tensor, dtype=dtype)
-                root_tensor = tf.cast(root_tensor, dtype=dtype)
-                broadcasted_tensor = hvd.broadcast(tensor, root_rank, process_set=this_set)
-            self.assertTrue(
-                self.evaluate(tf.reduce_all(tf.equal(
-                    tf.cast(root_tensor, tf.int32), tf.cast(broadcasted_tensor, tf.int32)))),
-                "hvd.broadcast produces incorrect broadcasted tensor")
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
-    def test_horovod_broadcast_gpu_process_sets(self):
-        """Test that the broadcast correctly broadcasts 1D, 2D, 3D tensors on GPU
-         if restricted to non-global process sets"""
-        # Only do this test if there are GPUs available.
-        if not tf.test.is_gpu_available(cuda_only=True):
-            self.skipTest("No GPUs available")
-
-        if int(os.environ.get('HOROVOD_MIXED_INSTALL', 0)):
-            # Skip if compiled with CUDA but without HOROVOD_GPU_OPERATIONS.
-            self.skipTest("Not compiled with HOROVOD_GPU_OPERATIONS")
-
-        hvd.init()
-        rank = hvd.rank()
-        local_rank = hvd.local_rank()
-        size = hvd.size()
-
-        if hvd.ccl_built():
-            self.skipTest("Multiple process sets currently do not support CCL.")
-
-        # This test does not apply if there is only one worker.
-        if size == 1:
-            self.skipTest("Only one worker available")
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        if rank in even_ranks:
-            set_size = len(even_ranks)
-            set_ranks = even_ranks
-            this_set = even_set
-        elif rank in odd_ranks:
-            set_size = len(odd_ranks)
-            set_ranks = odd_ranks
-            this_set = odd_set
-
-        dtypes = [tf.uint8, tf.int8, tf.uint16, tf.int16,
-                  tf.int32, tf.int64, tf.float16, tf.float32,
-                  tf.float64, tf.bool]
-        dims = [1, 2, 3]
-        root_ranks = list(set_ranks)
-        for dtype, dim, root_rank in itertools.product(dtypes, dims, root_ranks):
-            tensor = tf.ones([17] * dim) * rank
-            root_tensor = tf.ones([17] * dim) * root_rank
-            if dtype == tf.bool:
-                tensor = tensor % 2
-                root_tensor = root_tensor % 2
-            tensor = tf.cast(tensor, dtype=dtype)
-            root_tensor = tf.cast(root_tensor, dtype=dtype)
-            with tf.device("/gpu:%d" % local_rank):
-                broadcasted_tensor = hvd.broadcast(tensor, root_rank, process_set=this_set)
-            self.assertTrue(
-                self.evaluate(tf.reduce_all(tf.equal(
-                    tf.cast(root_tensor, tf.int32), tf.cast(broadcasted_tensor, tf.int32)))),
-                "hvd.broadcast produces incorrect broadcasted tensor")
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
-    def test_broadcast_variables_process_sets(self):
-        hvd.init()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        if hvd.ccl_built():
-            self.skipTest("Multiple process sets currently do not support CCL.")
-
-        # This test does not apply if there is only one worker.
-        if size == 1:
-            self.skipTest("Only one worker available")
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        if rank in even_ranks:
-            set_ranks = even_ranks
-            this_set = even_set
-        elif rank in odd_ranks:
-            set_ranks = odd_ranks
-            this_set = odd_set
-        root_rank = set_ranks[0]
-
-        with tf.device("/cpu:0"):
-            var = tf.Variable(initial_value=[rank], dtype=tf.int32)
-            if not hvd._executing_eagerly():
-                init = tf.compat.v1.global_variables_initializer()
-                self.evaluate(init)
-            self.evaluate(
-                hvd.broadcast_variables([var], root_rank=root_rank, process_set=this_set))
-            value = self.evaluate(var)
-        self.assertListEqual(list(value), [root_rank])
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
     def test_horovod_broadcast_error(self):
         """Test that the broadcast returns an error if any dimension besides
         the first is different among the tensors being broadcasted."""
@@ -2677,72 +1852,6 @@ class TensorFlowTests(tf.test.TestCase):
             self.assertLess(err, 0.00000001,
                             "gradient %s differs from expected %s, "
                             "error: %s" % (grad_out, expected, str(err)))
-
-    def test_horovod_broadcast_grad_cpu_process_sets(self):
-        """Test the correctness of the broadcast gradient on CPU if restricted to non-global process sets."""
-        hvd.init()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        if hvd.ccl_built():
-            self.skipTest("Multiple process sets currently do not support CCL.")
-
-        # This test does not apply if there is only one worker.
-        if size == 1:
-            self.skipTest("Only one worker available")
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        if rank in even_ranks:
-            set_size = len(even_ranks)
-            set_ranks = even_ranks
-            this_set = even_set
-        elif rank in odd_ranks:
-            set_size = len(odd_ranks)
-            set_ranks = odd_ranks
-            this_set = odd_set
-
-        # As of TensorFlow v1.9, gradients are not supported on
-        # integer tensors
-        dtypes = [tf.float32, tf.float64]
-        dims = [1, 2, 3]
-        root_ranks = list(set_ranks)
-        for dtype, dim, root_rank in itertools.product(dtypes, dims, root_ranks):
-            with tf.device("/cpu:0"):
-                if _executing_eagerly():
-                    tensor = self.tfe.Variable(tf.ones([5] * dim) * rank)
-                else:
-                    tensor = tf.ones([5] * dim) * rank
-                if dtype == tf.bool:
-                    tensor = tensor % 2
-                if _executing_eagerly():
-                    with tf.GradientTape() as tape:
-                        tensor = tf.cast(tensor, dtype=dtype)
-                        broadcasted_tensor = hvd.broadcast(tensor, root_rank,
-                                                           process_set=this_set)
-                    grad_out = tape.gradient(broadcasted_tensor, tensor)
-                else:
-                    tensor = tf.cast(tensor, dtype=dtype)
-                    broadcasted_tensor = hvd.broadcast(tensor, root_rank,
-                                                       process_set=this_set)
-                    grad_ys = tf.ones([5] * dim)
-                    grad = tf.gradients(broadcasted_tensor, tensor, grad_ys)[0]
-                    grad_out = self.evaluate(grad)
-
-            c = 1 if rank == root_rank else 0
-            expected = np.ones([5] * dim) * c
-            err = np.linalg.norm(expected - grad_out)
-            self.assertLess(err, 0.00000001,
-                            "gradient %s differs from expected %s, "
-                            "error: %s" % (grad_out, expected, str(err)))
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
 
     def test_horovod_alltoall_cpu(self):
         """Test that the alltoall correctly distributes 1D, 2D, and 3D tensors."""
@@ -3253,128 +2362,6 @@ class TensorFlowTests(tf.test.TestCase):
         self.assertSequenceEqual(self.evaluate(received_splits).tolist(), expected_rsplits,
                                  "hvd.alltoall returned incorrect received_splits")
 
-    def test_horovod_alltoall_cpu_process_sets(self):
-        """Test that the alltoall on restricted process sets correctly distributes 1D, 2D, and 3D tensors."""
-        hvd.init()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        if hvd.ccl_built():
-            self.skipTest("Multiple process sets currently do not support CCL.")
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        if rank in even_ranks:
-            set_size = len(even_ranks)
-            set_ranks = even_ranks
-        elif rank in odd_ranks:
-            set_size = len(odd_ranks)
-            set_ranks = odd_ranks
-
-        dtypes = self.filter_supported_types([tf.uint8, tf.int8, tf.uint16, tf.int16,
-                                              tf.int32, tf.int64, tf.float16, tf.float32,
-                                              tf.float64])
-        dims = [1, 2, 3]
-        for dtype, dim in itertools.product(dtypes, dims):
-            with tf.device("/cpu:0"):
-                vals = []
-                for i in set_ranks:
-                  vals += [i] * (rank+1)
-                tensor = tf.convert_to_tensor(vals, dtype=dtype)
-                for _ in range(dim - 1):
-                  tensor = tf.expand_dims(tensor, axis=1)
-                  tensor = tf.concat([tensor, tensor], axis=1)
-                splits = tf.convert_to_tensor([rank+1] * set_size, dtype=tf.int32)
-                if rank in even_ranks:
-                    collected, received_splits = hvd.alltoall(tensor, splits, process_set=even_set)
-                elif rank in odd_ranks:
-                    collected, received_splits = hvd.alltoall(tensor, splits, process_set=odd_set)
-
-                self.assertTrue(
-                    self.evaluate(tf.reduce_all(
-                        tf.equal(tf.cast(collected, tf.int32), rank))),
-                    "hvd.alltoall produces incorrect collected tensor")
-
-                self.assertTrue(
-                    self.evaluate(tf.equal(tf.size(collected), sum(rk + 1 for rk in set_ranks) * 2**(dim - 1))),
-                    "hvd.alltoall collected wrong number of values")
-
-                self.assertSequenceEqual(self.evaluate(received_splits).tolist(), [rk + 1 for rk in set_ranks],
-                                         "hvd.alltoall returned incorrect received_splits")
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
-    def test_horovod_alltoall_gpu_process_sets(self):
-        """Test that the GPU alltoall on restricted process sets correctly distributes 1D, 2D, and 3D tensors."""
-        if not tf.test.is_gpu_available(cuda_only=True):
-            self.skipTest("No GPUs available")
-
-        if int(os.environ.get('HOROVOD_MIXED_INSTALL', 0)):
-            # Skip if compiled with CUDA but without HOROVOD_GPU_OPERATIONS.
-            self.skipTest("Not compiled with HOROVOD_GPU_OPERATIONS")
-
-        # This test does not apply if NCCL version < 2.7.0
-        if hvd.nccl_built() and hvd.nccl_built() < 2700:
-            self.skipTest("NCCL-based Alltoall requires NCCL version >= 2.7.0.")
-
-        hvd.init()
-        rank = hvd.rank()
-        local_rank = hvd.local_rank()
-        size = hvd.size()
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        if rank in even_ranks:
-            set_size = len(even_ranks)
-            set_ranks = even_ranks
-        elif rank in odd_ranks:
-            set_size = len(odd_ranks)
-            set_ranks = odd_ranks
-
-        dtypes = [tf.uint8, tf.int8, tf.uint16, tf.int16,
-                  tf.int32, tf.int64, tf.float16, tf.float32,
-                  tf.float64]
-        dims = [1, 2, 3]
-        for dtype, dim in itertools.product(dtypes, dims):
-            with tf.device("/gpu:%s" % local_rank):
-                vals = []
-                for i in set_ranks:
-                  vals += [i] * (rank+1)
-                tensor = tf.convert_to_tensor(vals, dtype=dtype)
-                for _ in range(dim - 1):
-                  tensor = tf.expand_dims(tensor, axis=1)
-                  tensor = tf.concat([tensor, tensor], axis=1)
-                splits = tf.convert_to_tensor([rank+1] * set_size, dtype=tf.int32)
-                if rank in even_ranks:
-                    collected, received_splits = hvd.alltoall(tensor, splits, process_set=even_set)
-                elif rank in odd_ranks:
-                    collected, received_splits = hvd.alltoall(tensor, splits, process_set=odd_set)
-
-                self.assertTrue(
-                    self.evaluate(tf.reduce_all(
-                        tf.equal(tf.cast(collected, tf.int32), rank))),
-                    "hvd.alltoall produces incorrect collected tensor")
-
-                self.assertTrue(
-                    self.evaluate(tf.equal(tf.size(collected), sum(rk + 1 for rk in set_ranks) * 2**(dim - 1))),
-                    "hvd.alltoall collected wrong number of values")
-
-                self.assertSequenceEqual(self.evaluate(received_splits).tolist(), [rk + 1 for rk in set_ranks],
-                                         "hvd.alltoall returned incorrect received_splits")
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
-
     def test_horovod_alltoall_type_error(self):
         """Test that the alltoall returns an error if the tensor types differ
            across the processes."""
@@ -3639,68 +2626,6 @@ class TensorFlowTests(tf.test.TestCase):
                             "gradient %s differs from expected %s, "
                             "error: %s" % (grad_out, expected, str(err)))
 
-    def test_horovod_alltoall_grad_cpu_process_sets(self):
-        """Test the correctness of the alltoall gradient on CPU with restricted process sets."""
-        hvd.init()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        if hvd.ccl_built():
-            self.skipTest("Multiple process sets currently do not support CCL.")
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        if rank in even_ranks:
-            set_size = len(even_ranks)
-            this_set = even_set
-        elif rank in odd_ranks:
-            set_size = len(odd_ranks)
-            this_set = odd_set
-
-        # As of TensorFlow v1.9, gradients are not supported on
-        # integer tensors
-        dtypes = [tf.float32, tf.float64]
-        dims = [1, 2, 3]
-        for dtype, dim in itertools.product(dtypes, dims):
-            with tf.device("/cpu:0"):
-                vals = []
-                for i in range(set_size):
-                  vals += [i] * (rank+1)
-                tensor = tf.convert_to_tensor(vals, dtype=dtype)
-                for _ in range(dim - 1):
-                  tensor = tf.expand_dims(tensor, axis=1)
-                  tensor = tf.concat([tensor, tensor], axis=1)
-
-                if _executing_eagerly():
-                    tensor = self.tfe.Variable(tensor)
-                    splits = tf.convert_to_tensor([rank + 1] * set_size, dtype=tf.int32)
-                    with tf.GradientTape() as tape:
-                        collected, received_splits = hvd.alltoall(tensor, splits, process_set=this_set)
-                else:
-                    splits = tf.convert_to_tensor([rank + 1] * set_size, dtype=tf.int32)
-                    collected, received_splits = hvd.alltoall(tensor, splits, process_set=this_set)
-
-                grad_ys = tf.ones(tf.shape(collected))
-                if _executing_eagerly():
-                    grad_out = tape.gradient(collected, tensor, grad_ys)
-                else:
-                    grad = tf.gradients(collected, tensor, grad_ys)[0]
-                    grad_out = self.evaluate(grad)
-
-            expected = np.ones(tensor.get_shape().as_list())
-            err = np.linalg.norm(expected - grad_out)
-            self.assertLess(err, 0.00000001,
-                            "gradient %s differs from expected %s, "
-                            "error: %s" % (grad_out, expected, str(err)))
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
-
     def test_horovod_broadcast_eager_mode_error(self):
         """Test that tries to broadcast tensorflow global variables
         in eager execution mode. This call should raise a RuntimeError."""
@@ -3769,49 +2694,6 @@ class TensorFlowTests(tf.test.TestCase):
             obj = hvd.broadcast_object(obj, root_rank=0)
             self.assertDictEqual(obj, expected_obj)
 
-    def test_broadcast_object_process_sets(self):
-        """ This should best be tested with more than two Horovod processes """
-        hvd.init()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        if hvd.ccl_built():
-            self.skipTest("Multiple process sets currently do not support CCL.")
-
-        # This test does not apply if there is only one worker.
-        if size == 1:
-            self.skipTest("Only one worker available")
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-        if rank in even_ranks:
-            set_ranks = even_ranks
-            this_set = even_set
-        elif rank in odd_ranks:
-            set_ranks = odd_ranks
-            this_set = odd_set
-        root_rank = set_ranks[0]
-
-        with tf.device("/cpu:0"):
-            expected_even_obj = {
-                'even': 123,
-                0: [1, 2]
-            }
-            expected_odd_obj = {
-                'odd': 456,
-                1: [1, 2, 3, 4]
-            }
-            expected_obj = expected_even_obj if this_set == even_set else expected_odd_obj
-            obj = expected_obj if hvd.rank() == root_rank else {}
-
-            obj = hvd.broadcast_object(obj, root_rank=root_rank, process_set=this_set)
-            self.assertDictEqual(obj, expected_obj)
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
     def test_broadcast_object_fn(self):
         if hvd._executing_eagerly() or _IS_TF2:
             # Only for TF 1.0 in graph mode
@@ -3830,55 +2712,6 @@ class TensorFlowTests(tf.test.TestCase):
             obj = bcast(obj)
             self.assertDictEqual(obj, expected_obj)
 
-    def test_broadcast_object_fn_process_sets(self):
-        """ This should best be tested with more than two Horovod processes """
-        if hvd._executing_eagerly() or _IS_TF2:
-            # Only for TF 1.0 in graph mode
-            return
-
-        hvd.init()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        if hvd.ccl_built():
-            self.skipTest("Multiple process sets currently do not support CCL.")
-
-        # This test does not apply if there is only one worker.
-        if size == 1:
-            self.skipTest("Only one worker available")
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-        if rank in even_ranks:
-            set_ranks = even_ranks
-            this_set = even_set
-        elif rank in odd_ranks:
-            set_ranks = odd_ranks
-            this_set = odd_set
-        root_rank = set_ranks[0]
-
-        with tf.device("/cpu:0"):
-            expected_even_obj = {
-                'even': 123,
-                0: [1, 2]
-            }
-            expected_odd_obj = {
-                'odd': 456,
-                1: [1, 2, 3, 4]
-            }
-            expected_obj = expected_even_obj if this_set == even_set else expected_odd_obj
-            obj = expected_obj if hvd.rank() == root_rank else {}
-
-            bcast = hvd.broadcast_object_fn(root_rank=root_rank, process_set=this_set)
-            obj = bcast(obj)
-            self.assertDictEqual(obj, expected_obj)
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
-
     def test_allgather_object(self):
         hvd.init()
 
@@ -3895,50 +2728,6 @@ class TensorFlowTests(tf.test.TestCase):
 
             self.assertEqual(len(results), hvd.size())
             self.assertListEqual(results, expected)
-
-
-    def test_allgather_object_process_sets(self):
-        """ This should best be tested with more than two Horovod processes """
-        hvd.init()
-
-        rank = hvd.rank()
-        size = hvd.size()
-
-        if hvd.ccl_built():
-            self.skipTest("Multiple process sets currently do not support CCL.")
-
-        # This test does not apply if there is only one worker.
-        if size == 1:
-            self.skipTest("Only one worker available")
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-        if rank in even_ranks:
-            set_ranks = even_ranks
-            this_set = even_set
-        elif rank in odd_ranks:
-            set_ranks = odd_ranks
-            this_set = odd_set
-
-        with tf.device("/cpu:0"):
-            d = {'metric_val_1': hvd.rank()}
-            if this_set.rank() == 1:
-                d['metric_val_2'] = 42 if this_set == even_set else 23
-
-            results = hvd.allgather_object(d, process_set=this_set)
-
-            expected = [{'metric_val_1': i} for i in set_ranks]
-            if this_set.size() > 1:
-                expected[1] = {'metric_val_1': set_ranks[1],
-                               'metric_val_2': 42 if this_set == even_set else 23}
-
-            self.assertEqual(len(results), this_set.size())
-            self.assertListEqual(results, expected)
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
 
     def test_elastic_state(self):
         if not hvd._executing_eagerly() and _IS_TF2:
@@ -4262,164 +3051,6 @@ class TensorFlowTests(tf.test.TestCase):
                 self.assertEquals(computed_value, self._grad_agg_compute_expected_value(
                     backward_passes_per_step, idx))
 
-    def test_horovod_add_get_remove_process_set(self):
-        hvd.init()
-        size = hvd.size()
-
-        # This test does not apply if there is only one worker.
-        if size == 1:
-            self.skipTest("Only one worker available")
-
-        # Here we test some implementation details (numeric process set id values) using an internal function. We only
-        # test the concrete value 0 because IDs will be reassigned between eager and graph-mode test runs and may
-        # change.
-        ps = hvd.mpi_ops._basics._get_process_set_ids_and_ranks()
-        self.assertDictEqual(ps, {0: list(range(size))})
-
-        set1 = hvd.add_process_set([0])
-        set2 = hvd.add_process_set(range(1, size))
-
-        ps = hvd.mpi_ops._basics._get_process_set_ids_and_ranks()
-        self.assertDictEqual(ps, {0: list(range(size)),
-                                  set1.process_set_id: [0],
-                                  set2.process_set_id: list(range(1, size))})
-
-        # Ensure process set ids are equal across processes.
-        with tf.device("/cpu:0"):
-            for a_set in [set1, set2]:
-                ids_on_ranks = list(self.evaluate(hvd.allgather(tf.convert_to_tensor([a_set.process_set_id]))))
-                self.assertTrue(all(an_id == a_set.process_set_id for an_id in ids_on_ranks))
-
-        # Test stringification
-        self.assertListEqual([str(p) for p in [hvd.global_process_set, set1, set2]],
-                             [f"ProcessSet(process_set_id=0, ranks={list(range(size))}, mpi_comm=None)",
-                              f"ProcessSet(process_set_id={set1.process_set_id}, ranks=[0], mpi_comm=None)",
-                              f"ProcessSet(process_set_id={set2.process_set_id}, ranks={list(range(1, size))}, mpi_comm=None)",
-                              ])
-
-        old_id_of_set1 = set1.process_set_id
-        hvd.remove_process_set(set1)
-        self.assertIsNone(set1.process_set_id)  # invalidated
-
-        ps = hvd.mpi_ops._basics._get_process_set_ids_and_ranks()
-        self.assertDictEqual(ps, {0: list(range(size)),
-                                  set2.process_set_id: list(range(1, size))})
-
-        # test re-adding set1
-        hvd.add_process_set(set1)
-        ps = hvd.mpi_ops._basics._get_process_set_ids_and_ranks()
-        self.assertDictEqual(ps, {0: list(range(size)),
-                                  set1.process_set_id: [0],
-                                  set2.process_set_id: list(range(1, size))})
-        hvd.remove_process_set(set1)
-
-
-        if size > 2:
-            set3 = hvd.add_process_set([0, size - 1])
-            self.assertEqual(old_id_of_set1, set3.process_set_id) # id reuse
-        else:
-            with self.assertRaises(ValueError):  # duplicate of the global process set
-                set3 = hvd.add_process_set([0, size - 1])
-            set3 = hvd.global_process_set
-
-        with self.assertRaises(ValueError):  # duplicate of set2
-            set4 = hvd.add_process_set(range(size - 1, 0, -1))
-
-        with self.assertRaises(ValueError):  # duplicate of the global process set
-            set5 = hvd.add_process_set(range(0, size))
-
-        ps = hvd.mpi_ops._basics._get_process_set_ids_and_ranks()
-        if size > 2:
-            self.assertDictEqual(ps, {0: list(range(size)),
-                                      set2.process_set_id: list(range(1, size)),
-                                      set3.process_set_id: [0, size-1]})
-        else:
-            self.assertDictEqual(ps, {0: list(range(size)),
-                                      set2.process_set_id: list(range(1, size))})
-        hvd.remove_process_set(set2)
-        hvd.remove_process_set(set3)
-
-        ps = hvd.mpi_ops._basics._get_process_set_ids_and_ranks()
-        self.assertDictEqual(ps, {0: list(range(size))})
-
-        self.assertFalse(hvd.remove_process_set(hvd.global_process_set),
-                         "Removing the global process set should be impossible.")
-
-    def test_legacy_DistributedOptimizer_process_sets(self):
-        """ Note that this test makes the most sense when running with > 2 processes. """
-        if _executing_eagerly():
-            self.skipTest("Legacy Optimizers only support graph mode.")
-
-        resource_variables_by_default = tf.compat.v1.resource_variables_enabled()
-        tf.compat.v1.disable_resource_variables()
-
-        hvd.init()
-        size = hvd.size()
-
-        if size == 1:
-            self.skipTest("Only one worker available")
-
-        subset = hvd.add_process_set(range(0, size, 2))
-
-        with self.test_session() as sess:
-            class TestOptimizer(tf.compat.v1.train.Optimizer):
-                def __init__(self, *args, **kwargs):
-                    super().__init__(*args, **kwargs)
-                    self.variable = tf.compat.v1.get_variable("dummy_var", initializer=[0.0],
-                                                              use_resource=False)
-                def compute_gradients(self, *args, **kwargs):
-                    return [(tf.constant([float(hvd.rank())]), self.variable)]
-                def _apply_dense(self, grad, var):
-                    return var.assign_add(grad)
-
-            opt = TestOptimizer(name="test", use_locking=False)
-            opt = hvd.DistributedOptimizer(
-                optimizer=opt,
-                average_aggregated_gradients=True,
-                process_set=subset,
-            )
-
-            grads_and_vars = opt.compute_gradients()
-            update_op = opt.apply_gradients(grads_and_vars)
-            sess.run(tf.compat.v1.global_variables_initializer())
-            sess.run(update_op)
-
-            computed_value = sess.run(opt._optimizer.variable.read_value())[0]
-            if subset.included():
-                self.assertAlmostEqual(computed_value, sum(range(0, size, 2)) / subset.size())
-            else:
-                self.assertAlmostEqual(computed_value, float(hvd.rank()))
-
-        hvd.remove_process_set(subset)
-        if resource_variables_by_default:
-            tf.compat.v1.enable_resource_variables()
-
-    def test_distributed_gradient_tape_process_sets(self):
-        """ Note: test makes most sense with more than 2 nodes. """
-        hvd.init()
-        size = hvd.size()
-
-        if size == 1:
-            self.skipTest("Only one worker available")
-
-        subset = hvd.add_process_set(range(0, size, 2))
-
-        with tf.device("/cpu:0"):
-            x = tf.constant(float(hvd.rank()))
-            with tf.GradientTape() as g:
-                g.watch(x)
-                y = x * x
-            dg = hvd.DistributedGradientTape(g, process_set=subset)
-            dy_dx = dg.gradient(y, [x])
-        value, = self.evaluate(dy_dx)
-
-        if subset.included():
-            self.assertAlmostEqual(value, 2. * sum(subset.ranks) / subset.size())
-        else:
-            self.assertAlmostEqual(value, 2. * hvd.rank())
-
-        hvd.remove_process_set(subset)
-
     def test_horovod_reducescatter_cpu(self):
         """Test on CPU that the reducescatter correctly sums or averages and scatters 1D, 2D, 3D tensors."""
         if hvd.ccl_built():
@@ -4606,66 +3237,6 @@ class TensorFlowTests(tf.test.TestCase):
         # pprint(i)
         self.assertTrue(successful,
                         "hvd.reducescatter produces incorrect results")
-
-    def test_horovod_reducescatter_cpu_process_sets(self):
-        """Test on CPU that the reducescatter correctly sums or averages and scatters 1D, 2D, 3D tensors
-        if restricted to non-global process sets."""
-        if hvd.ccl_built():
-            self.skipTest("Reducescatter is not supported yet with oneCCL operations.")
-        if _is_mac and hvd.gloo_built() and not hvd.mpi_built():
-            self.skipTest("ReducescatterGloo is not supported on macOS")
-        hvd.init()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-        if rank in even_ranks:
-            this_set = even_set
-        else:
-            this_set = odd_set
-        process_set_size = this_set.size()
-        process_set_rank = this_set.rank()
-
-        dtypes = self.filter_supported_types([tf.int32, tf.int64, tf.float16, tf.float32, tf.float64])
-        dims = [1, 2, 3]
-        for red_op, dtype, dim in itertools.product([hvd.Sum, hvd.Average], dtypes, dims):
-            with tf.device("/cpu:0"):
-                even_rank_tensor = self.random_uniform([process_set_size * 4] * dim, -100, 100, dtype=dtype)
-                odd_rank_tensor = self.random_uniform([process_set_size * 4] * dim, -100, 100, dtype=dtype)
-                if rank in even_ranks:
-                    tensor = even_rank_tensor
-                else:
-                    tensor = odd_rank_tensor
-                reduced = hvd.reducescatter(tensor, op=red_op, process_set=this_set)
-            if red_op == hvd.Sum:
-                expected = tf.cast(tensor[process_set_rank * 4:(process_set_rank + 1) * 4] * process_set_size,
-                                   reduced.dtype)
-            elif red_op == hvd.Average:
-                expected = tf.cast(tensor[process_set_rank * 4:(process_set_rank + 1) * 4], reduced.dtype)
-            max_difference = tf.reduce_max(tf.abs(reduced - expected))
-
-            # Threshold for floating point equality depends on number of
-            # ranks, since we're comparing against precise multiplication.
-            if dtype == tf.float16:
-                threshold = .5
-            elif dtype in [tf.int32, tf.int64]:
-                threshold = 0
-            elif process_set_size < 10:
-                threshold = 1e-4
-            elif process_set_size < 15:
-                threshold = 5e-4
-            else:
-                break
-
-            diff = self.evaluate(max_difference)
-            self.assertTrue(diff <= threshold,
-                            "hvd.reducescatter produces incorrect results")
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
 
     def test_horovod_reducescatter_gpu(self):
         """Test that the reducescatter works on GPUs."""
@@ -4875,69 +3446,6 @@ class TensorFlowTests(tf.test.TestCase):
         self.assertTrue(succesful,
                         "hvd.reducescatter produces incorrect results")
 
-    def test_horovod_reducescatter_gpu_process_sets(self):
-        """Test that the reducescatter works on GPUs if restricted to non-global process sets."""
-        if not tf.test.is_gpu_available(cuda_only=True):
-            self.skipTest("No GPUs available")
-
-        if int(os.environ.get('HOROVOD_MIXED_INSTALL', 0)):
-            # Skip if compiled with CUDA but without HOROVOD_GPU_OPERATIONS.
-            self.skipTest("Not compiled with HOROVOD_GPU_OPERATIONS")
-
-        hvd.init()
-        local_rank = hvd.local_rank()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-        if rank in even_ranks:
-            this_set = even_set
-        else:
-            this_set = odd_set
-        process_set_size = this_set.size()
-        process_set_rank = this_set.rank()
-
-        dtypes = [tf.int32, tf.int64, tf.float16, tf.float32, tf.float64]
-        dims = [1, 2, 3]
-        for red_op, dtype, dim in itertools.product([hvd.Sum, hvd.Average], dtypes, dims):
-            with tf.device("/gpu:%d" % local_rank):
-                even_rank_tensor = self.random_uniform([process_set_size * 4] * dim, -100, 100, dtype=dtype)
-                odd_rank_tensor = self.random_uniform([process_set_size * 4] * dim, -100, 100, dtype=dtype)
-                if rank in even_ranks:
-                    tensor = even_rank_tensor
-                else:
-                    tensor = odd_rank_tensor
-                reduced = hvd.reducescatter(tensor, op=red_op, process_set=this_set)
-            if red_op == hvd.Sum:
-                expected = tf.cast(tensor[process_set_rank * 4:(process_set_rank + 1) * 4] * process_set_size,
-                                   reduced.dtype)
-            elif red_op == hvd.Average:
-                expected = tf.cast(tensor[process_set_rank * 4:(process_set_rank + 1) * 4], reduced.dtype)
-            max_difference = tf.reduce_max(tf.abs(reduced - expected))
-
-            # Threshold for floating point equality depends on number of
-            # ranks, since we're comparing against precise multiplication.
-            if dtype == tf.float16:
-                threshold = .5
-            elif dtype in [tf.int32, tf.int64]:
-                threshold = 0
-            elif process_set_size < 10:
-                threshold = 1e-4
-            elif process_set_size < 15:
-                threshold = 5e-4
-            else:
-                return
-
-            diff = self.evaluate(max_difference)
-            self.assertTrue(diff <= threshold,
-                            "hvd.reducescatter on GPU produces incorrect results")
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
-
     def test_horovod_reducescatter_error(self):
         """Test that the reducescatter raises an error if different ranks try to
         send tensors of different rank or dimension."""
@@ -5028,76 +3536,6 @@ class TensorFlowTests(tf.test.TestCase):
             self.assertLess(err, 0.00000001,
                             "gradient %s differs from expected %s, "
                             "error: %s" % (grad_out, expected, str(err)))
-
-    def test_horovod_reducescatter_grad_cpu_process_sets(self):
-        """Test the correctness of the reducescatter gradient on CPU if restricted to non-global process sets."""
-        if hvd.ccl_built():
-            self.skipTest("Reducescatter is not supported yet with oneCCL operations.")
-        if _is_mac and hvd.gloo_built() and not hvd.mpi_built():
-            self.skipTest("ReducescatterGloo is not supported on macOS")
-        hvd.init()
-        rank = hvd.rank()
-        size = hvd.size()
-
-        even_ranks = [rk for rk in range(0, size) if rk % 2 == 0]
-        odd_ranks = [rk for rk in range(0, size) if rk % 2 == 1]
-
-        even_set = hvd.add_process_set(even_ranks)
-        odd_set = hvd.add_process_set(odd_ranks)
-
-        # As of TensorFlow v1.9, gradients are not supported on
-        # integer tensors
-        dtypes = [tf.float16, tf.float32, tf.float64]
-        dims = [1, 2, 3]
-        for red_op, dtype, dim in itertools.product([hvd.Sum, hvd.Average], dtypes, dims):
-            with tf.device("/cpu:0"):
-                if _executing_eagerly():
-                    even_rank_tensor = self.tfe.Variable(self.random_uniform(
-                        [len(even_ranks) * 4] * dim, -100, 100, dtype=dtype))
-                    odd_rank_tensor = self.tfe.Variable(self.random_uniform(
-                        [len(odd_ranks) * 4] * dim, -100, 100, dtype=dtype))
-                    with tf.GradientTape() as tape:
-                        if rank in even_ranks:
-                            reduced = hvd.reducescatter(even_rank_tensor, op=red_op,
-                                                       process_set=even_set)
-                        elif rank in odd_ranks:
-                            reduced = hvd.reducescatter(odd_rank_tensor, op=red_op,
-                                                       process_set=odd_set)
-                else:
-                    even_rank_tensor = self.random_uniform([len(even_ranks) * 4] * dim, -100, 100, dtype=dtype)
-                    odd_rank_tensor = self.random_uniform([len(odd_ranks) * 4] * dim, -100, 100, dtype=dtype)
-                    if rank in even_ranks:
-                        reduced = hvd.reducescatter(even_rank_tensor, op=red_op,
-                                                   process_set=even_set)
-                    elif rank in odd_ranks:
-                        reduced = hvd.reducescatter(odd_rank_tensor, op=red_op,
-                                                   process_set=odd_set)
-
-                if rank in even_ranks:
-                    tensor = even_rank_tensor
-                    set_size = len(even_ranks)
-                elif rank in odd_ranks:
-                    tensor = odd_rank_tensor
-                    set_size = len(odd_ranks)
-
-                grad_ys = tf.ones([4] + [set_size * 4] * (dim - 1), dtype=tensor.dtype)
-                if _executing_eagerly():
-                    grad_out = tape.gradient(tf.cast(reduced, dtype=tensor.dtype), tensor, grad_ys)
-                else:
-                    grad = tf.gradients(tf.cast(reduced, dtype=tensor.dtype), tensor, grad_ys)[0]
-                    grad_out = self.evaluate(grad)
-
-            if red_op == hvd.Sum:
-                expected = np.ones([set_size * 4] * dim) * set_size
-            elif red_op == hvd.Average:
-                expected = np.ones([set_size * 4] * dim)
-            err = np.linalg.norm(expected - grad_out)
-            self.assertLess(err, 0.00000001,
-                            "gradient %s differs from expected %s, "
-                            "error: %s" % (grad_out, expected, str(err)))
-
-        hvd.remove_process_set(odd_set)
-        hvd.remove_process_set(even_set)
 
     def test_horovod_reducescatter_grad_gpu(self):
         """Test the correctness of the reducescatter gradient on GPU."""
